feat: rename "fixture" classes to "builder" classes for consistency
