+++ index.ts
@@ -1,14 +1,14 @@
-import Builder, { Options } from "./lib/builder";
 import Input from "./lib/input";
+import Output, { Options } from "./lib/output";
 
-export { default as FixtureBuilder } from "./lib/fixture";
-export { default as ApplicationFixture } from "./lib/fixtures/application";
-export { default as AddonFixture } from "./lib/fixtures/addon";
+export { default as FixtureBuilder } from "./lib/builder";
+export { default as ApplicationBuilder } from "./lib/builders/application";
+export { default as AddonBuilder } from "./lib/builders/addon";
 
 export async function createTempDir() {
   return await Input.create();
 }
 
 export function createBuilder(path: string, options?: Options) {
-  return new Builder(path, options);
+  return new Output(path, options);
 }
