+++ lib/builder.ts
@@ -1,62 +1,48 @@
-import { Changes, createBuilder, Output, Tree } from "broccoli-test-helper";
-import { join } from "path";
+function writeFile(filename: string, contents: string, receiver: object): void {
+  const pathParts = filename.split("/");
+  const directories = pathParts.length > 1 ? pathParts.slice(0, -1) : [];
+  const file = pathParts.slice(-1)[0];
 
-// PRIVATE API ACCESS
-const EmberApp: any = require("ember-cli/lib/broccoli/ember-app");
-const _resetTreeCache: any = require("ember-cli/lib/models/addon")._resetTreeCache;
+  let current: any = receiver;
+  for (const part of directories) {
+    current[part] = current[part] || {};
+    current = current[part];
+  }
 
-export interface Options { workingDir: string; }
+  current[file] = contents;
+}
 
 export default class Builder {
-  public path: string;
+  public files: Map<string, string> = new Map();
 
-  public options: Options;
-
-  private wrappedBuilder!: Output;
-
-  constructor(path: string, options?: Options) {
-    this.path = path;
-    this.options = Object.assign({
-      workingDir: "application"
-    }, options);
-  }
-
-  get builder(): Output {
-    if (this.wrappedBuilder) {
-      return this.wrappedBuilder;
+  public application(app: Map<string, string>): this {
+    for (const [filename, contents] of app) {
+      this.files.set(filename, contents);
     }
 
-    const cwd = process.cwd();
-    process.chdir(join(this.path, this.options.workingDir));
+    return this;
+  }
 
-    try {
-      const emberApp = new EmberApp();
-      const wrappedBuilder = createBuilder(emberApp.toTree([]));
-      this.wrappedBuilder = wrappedBuilder;
-    } finally {
-      process.chdir(cwd);
+  public addon(addon: Map<string, string>): this {
+    for (const [filename, contents] of addon) {
+      this.files.set(filename, contents);
     }
 
-    return this.wrappedBuilder;
+    return this;
   }
 
-  public build(): Promise<void> {
-    try {
-      return this.builder.build();
-    } finally {
-      _resetTreeCache();
+  public file(name: string, contents: string): this {
+    this.files.set(name, contents);
+    return this;
+  }
+
+  public build(): object {
+    const result = {};
+
+    for (const [name, contents] of this.files) {
+      writeFile(name, contents, result);
     }
-  }
 
-  public changes(): Changes {
-    return this.builder.changes();
-  }
-
-  public dispose(): Promise<void> {
-    return this.builder.dispose();
-  }
-
-  public read(from?: string): Tree {
-    return this.builder.read(from);
+    return result;
   }
 }
